requirements.txt
deepface==0.0.83
opencv-python==4.8.1.78
pillow==10.2.0
numpy==1.24.4
tensorflow==2.10.1
keras==2.10.0
matplotlib==3.7.1
gdown==4.7.1
pandas==1.5.3
cmake==3.26.4
dlib==19.24.0

